+++ avalanche/benchmarks/scenarios/__init__.py
@@ -4,7 +4,7 @@ from avalanche.benchmarks.scenarios.deprecated.classification_scenario import *
 from avalanche.benchmarks.scenarios.deprecated.new_classes import *
 from avalanche.benchmarks.scenarios.deprecated.new_instances import *
 
-from .experience_decorators import *
+from .task_aware import *
 from .dataset_scenario import *
 from .exmodel_scenario import *
-from .online_scenario import *
+from .online import *
