+++ tests/benchmarks/scenarios/test_experience_decorators.py
@@ -3,7 +3,8 @@ import unittest
 import torch
 
 from avalanche.benchmarks import benchmark_from_datasets, EagerCLStream, DatasetExperience, CLScenario
-from avalanche.benchmarks.scenarios.experience_decorators import with_classes_timeline, with_task_labels
+from avalanche.benchmarks.scenarios.task_aware import with_task_labels
+from avalanche.benchmarks.scenarios.supervised import with_classes_timeline
 from avalanche.benchmarks.utils import make_tensor_classification_dataset
