+++ tests/benchmarks/scenarios/test_rl_scenario.py
@@ -3,7 +3,7 @@ import numpy as np
 
 try:
     import gym
-    from avalanche.benchmarks.scenarios.rl_scenario import (
+    from avalanche.benchmarks.scenarios.reinforcement_learning import (
         RLScenario,
         RLExperience,
     )
